+++ src/Shrun/Configuration/Data/CommonLogging.hs
@@ -19,7 +19,6 @@ module Shrun.Configuration.Data.CommonLogging
 where
 
 import Shrun.Configuration.Data.CommonLogging.KeyHideSwitch (KeyHideSwitch)
-import Shrun.Configuration.Data.CommonLogging.TimerFormat (TimerFormat)
 import Shrun.Configuration.Data.ConfigPhase
   ( ConfigPhase
       ( ConfigPhaseArgs,
@@ -35,11 +34,9 @@ import Shrun.Prelude
 
 -- | Holds command logging config.
 type CommonLoggingP :: ConfigPhase -> Type
-data CommonLoggingP p = MkCommonLoggingP
+newtype CommonLoggingP p = MkCommonLoggingP
   { -- | Whether to display command by (key) name or command.
-    keyHide :: ConfigPhaseF p KeyHideSwitch,
-    -- | How to format the timer.
-    timerFormat :: ConfigPhaseF p TimerFormat
+    keyHide :: ConfigPhaseF p KeyHideSwitch
   }
 
 makeFieldLabelsNoPrefix ''CommonLoggingP
@@ -72,9 +69,7 @@ mergeCommonLogging ::
 mergeCommonLogging args mToml =
   MkCommonLoggingP
     { keyHide =
-        (args ^. #keyHide) <>?. (toml ^. #keyHide),
-      timerFormat =
-        (args ^. #timerFormat) <>?. (toml ^. #timerFormat)
+        (args ^. #keyHide) <>?. (toml ^. #keyHide)
     }
   where
     toml = fromMaybe defaultToml mToml
@@ -83,32 +78,25 @@ instance DecodeTOML CommonLoggingToml where
   tomlDecoder =
     MkCommonLoggingP
       <$> decodeKeyHideSwitch
-      <*> decodeCmdLineTrunc
 
 decodeKeyHideSwitch :: Decoder (Maybe KeyHideSwitch)
 decodeKeyHideSwitch = getFieldOptWith tomlDecoder "key-hide"
-
-decodeCmdLineTrunc :: Decoder (Maybe TimerFormat)
-decodeCmdLineTrunc = getFieldOptWith tomlDecoder "timer-format"
 
 -- | Creates env version from merged.
 toEnv :: CommonLoggingMerged -> CommonLoggingEnv
 toEnv merged =
   MkCommonLoggingP
-    { keyHide = merged ^. #keyHide,
-      timerFormat = merged ^. #timerFormat
+    { keyHide = merged ^. #keyHide
     }
 
 defaultToml :: CommonLoggingToml
 defaultToml =
   MkCommonLoggingP
-    { keyHide = Nothing,
-      timerFormat = Nothing
+    { keyHide = Nothing
     }
 
 defaultMerged :: CommonLoggingMerged
 defaultMerged =
   MkCommonLoggingP
-    { keyHide = def,
-      timerFormat = def
+    { keyHide = def
     }