+++ test/functional/Functional/Examples/CommonLogging.hs
@@ -11,11 +11,7 @@ tests =
   testGroup
     "CommonLogging"
     [ keyHideOn,
-      keyHideOff,
-      timerFormatDigitalCompact,
-      timerFormatDigitalFull,
-      timerFormatProseCompact,
-      timerFormatProseFull
+      keyHideOff
     ]
 
 keyHideOn :: TestTree
@@ -55,68 +51,4 @@ keyHideOff =
     unexpected =
       [ withCommandPrefix "echo hi && sleep 2" "",
         withSuccessPrefix "echo hi && sleep 2"
-      ]
-
-timerFormatDigitalCompact :: TestTree
-timerFormatDigitalCompact =
-  testCase "Runs timer format with digital_compact" $ do
-    results <- fmap MkResultText <$> (readIORef =<< run args)
-    V.verifyExpected results expected
-  where
-    args =
-      withBaseArgs
-        [ "--log-timer-format",
-          "digital_compact",
-          "sleep 2"
-        ]
-    expected =
-      [ withTimerPrefix "01"
-      ]
-
-timerFormatDigitalFull :: TestTree
-timerFormatDigitalFull =
-  testCase "Runs timer format with digital_full" $ do
-    results <- fmap MkResultText <$> (readIORef =<< run args)
-    V.verifyExpected results expected
-  where
-    args =
-      withBaseArgs
-        [ "--log-timer-format",
-          "digital_full",
-          "sleep 2"
-        ]
-    expected =
-      [ withTimerPrefix "00:00:00:01"
-      ]
-
-timerFormatProseCompact :: TestTree
-timerFormatProseCompact =
-  testCase "Runs timer format with prose_compact" $ do
-    results <- fmap MkResultText <$> (readIORef =<< run args)
-    V.verifyExpected results expected
-  where
-    args =
-      withBaseArgs
-        [ "--log-timer-format",
-          "prose_compact",
-          "sleep 2"
-        ]
-    expected =
-      [ withTimerPrefix "1 second"
-      ]
-
-timerFormatProseFull :: TestTree
-timerFormatProseFull =
-  testCase "Runs timer format with prose_full" $ do
-    results <- fmap MkResultText <$> (readIORef =<< run args)
-    V.verifyExpected results expected
-  where
-    args =
-      withBaseArgs
-        [ "--log-timer-format",
-          "prose_full",
-          "sleep 2"
-        ]
-    expected =
-      [ withTimerPrefix "0 days, 0 hours, 0 minutes, 1 second"
       ]