+++ test/functional/Functional/Examples/ConsoleLogging.hs
@@ -17,7 +17,11 @@ tests =
       commandLogLineTruncN,
       stripControlAll,
       stripControlNone,
-      stripControlSmart
+      stripControlSmart,
+      timerFormatDigitalCompact,
+      timerFormatDigitalFull,
+      timerFormatProseCompact,
+      timerFormatProseFull
     ]
 
 commandLogOn :: TestTree
@@ -149,4 +153,68 @@ stripControlSmart = testCase "Runs --console-log-strip-control smart example" $ 
         ]
     expected =
       [ withCommandPrefix "printf ..." "foo \ESC[35m hello  bye"
+      ]
+
+timerFormatDigitalCompact :: TestTree
+timerFormatDigitalCompact =
+  testCase "Runs timer format with digital_compact" $ do
+    results <- fmap MkResultText <$> (readIORef =<< run args)
+    V.verifyExpected results expected
+  where
+    args =
+      withBaseArgs
+        [ "--console-log-timer-format",
+          "digital_compact",
+          "sleep 2"
+        ]
+    expected =
+      [ withTimerPrefix "01"
+      ]
+
+timerFormatDigitalFull :: TestTree
+timerFormatDigitalFull =
+  testCase "Runs timer format with digital_full" $ do
+    results <- fmap MkResultText <$> (readIORef =<< run args)
+    V.verifyExpected results expected
+  where
+    args =
+      withBaseArgs
+        [ "--console-log-timer-format",
+          "digital_full",
+          "sleep 2"
+        ]
+    expected =
+      [ withTimerPrefix "00:00:00:01"
+      ]
+
+timerFormatProseCompact :: TestTree
+timerFormatProseCompact =
+  testCase "Runs timer format with prose_compact" $ do
+    results <- fmap MkResultText <$> (readIORef =<< run args)
+    V.verifyExpected results expected
+  where
+    args =
+      withBaseArgs
+        [ "--console-log-timer-format",
+          "prose_compact",
+          "sleep 2"
+        ]
+    expected =
+      [ withTimerPrefix "1 second"
+      ]
+
+timerFormatProseFull :: TestTree
+timerFormatProseFull =
+  testCase "Runs timer format with prose_full" $ do
+    results <- fmap MkResultText <$> (readIORef =<< run args)
+    V.verifyExpected results expected
+  where
+    args =
+      withBaseArgs
+        [ "--console-log-timer-format",
+          "prose_full",
+          "sleep 2"
+        ]
+    expected =
+      [ withTimerPrefix "0 days, 0 hours, 0 minutes, 1 second"
       ]